restore: Fix pipes restore if both pipe ends are used by former process
In case if the former process used both pipe ends we fall
into endless loop since the 'users' never reaches 1. So
to resolve this during the pipe collectioning we check if
both pipe ends were used by a former process and if so --
we connect first pipe end immediately in create_pipe and
connect the rest via  attach_pipe.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>