parasite: Remove extra on-stack regsets
In fact all we need for parasitic syscall is -- original regs and some
regs with syscall args to work on. The former one is obtained in the
infect/cure code and the latter one is located (and copied from orig)
nearby.

That said, we can simplify all the regs passing-down code by removing
intermediate instances and utilizing the top-level reg-s for this.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>