parasite.c: use .space not .skip for inline asm
In the beginning it was the Word, and the Word was .space.
Later, .skip was introduced as a synonym for .space,
initially only for 680x0 platform.

But the sole reason for this patch is that I use clang/llvm
to check the code and unfortunately it does not understand .skip,
printing long ugly warnings I'd like to avoid.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>